+++ src/ts/indium.ts
@@ -1,2 +1,12 @@
-// export { Zone } from './zone';
-export { Screen } from './zones/screen';
+export { TouchScreen }  from './touchscreen';
+
+export { Circle }       from './zones/circle';
+export { Rectangle }    from './zones/rectangle';
+export { Top }          from './zones/top';
+export { Right }        from './zones/right';
+export { Bottom }       from './zones/bottom';
+export { Left }         from './zones/left';
+export { TopLeft }      from './zones/topleft';
+export { TopRight }     from './zones/topright';
+export { BottomLeft }   from './zones/bottomleft';
+export { BottomRight }  from './zones/bottomright';