+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-
-Version 0.1.0 (May 1st 2016)
+Version 0.2.0 (May 5th 2019)
 -----------------------------
- * initial version
+ * Added zones : Top, Left, Right, Bottom
+
+Version 0.1.0 (May 1st 2019)
+-----------------------------
+ * Initial version
+ * gestures : TouchStart, TouchMove, TouchEnd, TouchCancel